+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Implement `Create mediated request` form with basic functionality. Refs UIREQMED-27.
 * Add Mediated request status and Mediated request level filters for Mediated requests actions page. Refs UIREQMED-25.
 * Add Search for Mediated requests actions page. Refs UIREQMED-26.
+* Add Actions button and dropdown in Mediated requests UI. Refs UIREQMED-6.
 
 ## 1.0.0
 * New app created with stripes-cli. Updated module after created with stripes-cli. Refs UIREQMED-1.